# docs/openapi.yaml
openapi: 3.0.3
info:
  title: InvestIQ API
  description: |
    InvestIQ is a comprehensive algorithmic trading platform with multi-engine analysis,
    backtesting, portfolio management, and autonomous trading capabilities.

    ## Authentication
    All endpoints (except `/health` and `/metrics`) require authentication via the `X-API-Key` header or Bearer token.

    ## Live Trading Protection
    Broker write endpoints (POST /api/broker/execute, DELETE /api/broker/positions/:symbol, POST /api/broker/orders/:id/cancel)
    require an additional `X-Live-Trading-Key` header for safety.

    ## Response Format
    All API responses follow a standard wrapper:
    ```json
    {
      "success": true,
      "data": { ... },
      "error": null
    }
    ```
  version: 1.0.0
  contact:
    name: InvestIQ Support
    url: https://github.com/yourusername/invest-iq
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.investiq.io
    description: Production server
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/backtest/run:
    post:
      tags: [Backtest]
      summary: Run a custom backtest with full configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunBacktestRequest'
      responses:
        '200':
          description: Backtest completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /api/backtest/results/{id}:
    get:
      tags: [Backtest]
      summary: Get a specific backtest result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Backtest result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      tags: [Backtest]
      summary: Delete a backtest result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Backtest deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/backtest/results/{id}/trades:
    get:
      tags: [Backtest]
      summary: Get trades for a backtest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of trades
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/backtest/results/{id}/monte-carlo:
    get:
      tags: [Backtest]
      summary: Run Monte Carlo simulation on backtest trades
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: simulations
          in: query
          schema:
            type: integer
            default: 1000
            minimum: 100
            maximum: 10000
      responses:
        '200':
          description: Monte Carlo result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/backtest/strategy/{name}:
    get:
      tags: [Backtest]
      summary: Get backtests by strategy name
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of backtest results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/backtest/walk-forward:
    post:
      tags: [Backtest]
      summary: Run walk-forward validation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunBacktestRequest'
      responses:
        '200':
          description: Walk-forward result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # ============================================================================
  # Portfolio
  # ============================================================================
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication. Can also use Bearer token in Authorization header.

    LiveTradingAuth:
      type: apiKey
      in: header
      name: X-Live-Trading-Key
      description: Additional key required for live trading write operations.

    AdminAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Admin-level API key required for privileged operations.
  schemas:
    ApiResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        data:
          type: object
          description: Response data (structure varies by endpoint)
        error:
          type: string
          nullable: true
...
    RunBacktestRequest:
      type: object
      required: [strategy_name, symbols, start_date, end_date, initial_capital, position_size_percent, confidence_threshold]
      properties:
        strategy_name:
          type: string
        symbols:
          type: array
          items:
            type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        initial_capital:
          type: number
          example: 100000
        position_size_percent:
          type: number
          example: 95
        stop_loss_percent:
          type: number
        take_profit_percent:
          type: number
        confidence_threshold:
          type: number
          example: 0.5
        allocation_strategy:
          type: string
          enum: [equal_weight, custom]
        symbol_weights:
          type: object
          additionalProperties:
            type: number
        rebalance_interval_days:
          type: integer
...